* Fixes problem with webpack mangling constructor.name (should not rely on constructor.name and use constructor instead) * Fix demo game tests that were relying on a class property collision between .list (now scrollableList) and Phaser's .list
