+++ test/components/shop/shop-confirm-screen.test.js
@@ -10,6 +10,7 @@ import * as title from "../../../src/core/titles.js";
 import * as balance from "../../../src/components/shop/balance-ui.js";
 import * as confirm from "../../../src/components/shop/confirm.js";
 import { ShopConfirm } from "../../../src/components/shop/shop-confirm-screen.js";
+import { initResizers } from "../../../src/components/shop/backgrounds.js";
 
 jest.mock("../../../src/core/titles.js");
 jest.mock("../../../src/components/shop/balance-ui.js");
@@ -52,6 +53,13 @@ describe("Shop Confirm Screen", () => {
             transient: { shop: { title: "shop", item: "item" } },
             config: { "shop-confirm": { shopConfig: mockShopConfig } },
         };
+
+        global.RexPlugins = {
+            GameObjects: {
+                NinePatch: jest.fn(),
+            },
+        };
+        initResizers();
     });
     afterEach(() => jest.clearAllMocks());
