+++ test/components/shop/shop-list-screen.test.js
@@ -11,6 +11,7 @@ import * as balance from "../../../src/components/shop/balance-ui.js";
 import * as list from "../../../src/components/shop/scrollable-list/scrollable-list.js";
 import * as gmi from "../../../src/core/gmi/gmi.js";
 import { ShopList } from "../../../src/components/shop/shop-list-screen.js";
+import { initResizers } from "../../../src/components/shop/backgrounds.js";
 
 jest.mock("../../../src/core/titles.js");
 jest.mock("../../../src/components/shop/balance-ui.js");
@@ -56,6 +57,13 @@ describe("Shop List Screen", () => {
             transient: { shop: { mode: "shop", config: { balance: { value: { key: "balance" } } } } },
             config: { "shop-menu": { shopConfig: mockShopConfig } },
         };
+
+        global.RexPlugins = {
+            GameObjects: {
+                NinePatch: jest.fn(),
+            },
+        };
+        initResizers();
     });
     afterEach(() => jest.clearAllMocks());
 
@@ -114,6 +122,7 @@ describe("Shop List Screen", () => {
 
     test("creates list and adds reference to screen on create", () => {
         shopList.create();
+
         expect(list.ScrollableList).toHaveBeenCalledWith(
             shopList,
             shopList._data.transient.shop.mode,
